backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-multipart
python-dotenv

# Test dependencies
pytest
httpx